requirements.txt
chainlit==0.4.101
langchain==0.0.217
python-dotenv==1.0.0
